netfilter: nf_tables: add in-kernel only query that will return expired/dead elements
Extend the rhashtable backend so that nf_tables core can lookup a dead
or expired entry.

As-is, a set lookup will never return a matching entry that has expired,
which is the desired behaviour.

This has to be toggleable, else a dead element can shadow an alive one,
or vice versa.

Add internal NFT_SET_ELEM_GET_DEAD: if set, return a dead/expired
element, or null.
This is used in a follow patch that will switch the current gc
collection scheme from 'collect expired elements' to 'collect
the keys of expired elements'.

The difference is that in the former case, extra assurances are needed
to prevent the collected pointers from being free'd, or (this is the
current solution) a way to detect when the element pointers might have
gone stale.

With keys, we can store a full copy, and collect keys of elements that
have been expired.

We can then query them later and no longer need to worry about removal
of the element after it was picked up by gc, but beofre the gc worker
ran.

Signed-off-by: Florian Westphal <fw@strlen.de>