netfilter: nf_tables: prepare for key-based deletion from workqueue
At the moment async gc stores pointers to the to-be-removed
elements, async workqueue removes and them.

This works but has a few drawbacks:
Care must be taken so no pointers is posted twice, else we get
double-free.
This makes it necessary to skip a gc cycle if the work queue
is still scheduled, even though the given set has no pending
gc deletions.

Because pointers are used, we also have to consider userspace
transactions flushing/altering the set while gc is pending,
for this reason we have to back out if userspace made some changes,
even though the set might not have been affected at all and all
queued element pointers are still valid.

Now that rhashtable backend is the only remaining async gc
worker, lets simplify this.

Instead of storing pointers, store a copy of the element keys
that have expired.

Then, from work queue with transaction mutex held, query
the set for the given key.

If its found and expired, remove it.

IOW, the idea is to collect *candidates* to examine
for expiration, not element pointers.

If the element got removed, element lookup will return
no match and key is skipped, same if the key isn't expired
(i.e. userspace flushed and repopulated the set with
 identical, active key).

This is an intermediate step:
We still store pointers and use them, but also store the keys.
We then validate that lookup does returns the expected pointer.

Use WARN()s, they are not expected to trigger because the
gc sequence checks are still in place.

Next patch can remove the gc sequence counter and only rely
on the key lookup.

Signed-off-by: Florian Westphal <fw@strlen.de>